Fallback to get_common_ancestor_commit_hash for manual GL pipelines
Workaround/fallback for undocumented GitLab behaviour: if a pipeline is
manually run against any branch (i.e. it is not triggered by a commit or
by a MR or PR), CI_COMMIT_BEFORE_SHA is set to a string of 40 zeros. In
such case, to avoid that commit range check fails with error, we need to
fallback to get_common_ancestor_commit_hash

Signed-off-by: Alberto Pianon <alberto@pianon.eu>