Fix GitLab pipelines on default branch with no new commits
In GitLab, when a pipeline is launched but there are no new commits
(eg. when a pipeline is manually launched via GL webUI),
CI_COMMIT_BEFORE_SHA is set to a string of 40 zeros. This fix expressly
handles this case, to avoid internal python errors and to correctly skip
DCO check when it's not needed.

Signed-off-by: Alberto Pianon <alberto@pianon.eu>